Slettet DeserializeTests (alle i samme fil nå) og gitt navn på metode i AuthTests
